+++ test/e2e/tests/dapp-interactions/multichain-accounts/common.ts
@@ -1,5 +1,6 @@
 import { Driver } from '../../../webdriver/driver';
 import { Mockttp } from 'mockttp';
+import { strict as assert } from 'assert';
 
 import {
   mockGetMinimumBalanceForRentExemption,
@@ -17,13 +18,25 @@ import {
   mockGetAccountInfoDevnet,
 } from '../../solana/common-solana';
 import FixtureBuilder from '../../../fixture-builder';
-import { DEFAULT_MULTICHAIN_TEST_DAPP_FIXTURE_OPTIONS } from '../../../flask/multichain-api/testHelpers';
-import { withFixtures } from '../../../helpers';
+import {
+  DEFAULT_MULTICHAIN_TEST_DAPP_FIXTURE_OPTIONS,
+  getExpectedSessionScope,
+} from '../../../flask/multichain-api/testHelpers';
+import {
+  DAPP_HOST_ADDRESS,
+  largeDelayMs,
+  WINDOW_TITLES,
+  withFixtures,
+} from '../../../helpers';
 import { loginWithoutBalanceValidation } from '../../../page-objects/flows/login.flow';
 import { mockMultichainAccountsFeatureFlagStateTwo } from '../../multichain-accounts/common';
 import HeaderNavbar from '../../../page-objects/pages/header-navbar';
 import { SECOND_TEST_E2E_SRP } from '../../../flask/multi-srp/common-multi-srp';
 import HomePage from '../../../page-objects/pages/home/homepage';
+import ConnectAccountConfirmation from '../../../page-objects/pages/confirmations/redesign/connect-account-confirmation';
+import PermissionListPage from '../../../page-objects/pages/permission/permission-list-page';
+import SitePermissionPage from '../../../page-objects/pages/permission/site-permission-page';
+import TestDappMultichain from '../../../page-objects/pages/test-dapp-multichain';
 
 export const SECOND_MULTICHAIN_ACCOUNT_IN_SECOND_TEST_E2E = {
   evm: '0x695d565c9dbcddd65845162423bc2ad3700081ff',
@@ -102,3 +115,141 @@ export const withMultichainAccountsAndDappConnectionFixture = async (
     },
   );
 };
+
+export function validateSessionScopes(
+  sessionResult: any,
+  multichainAccount: {
+    evm: string;
+    solana: string;
+  },
+  evmScopes: string[] = EXPECTED_EVM_SCOPES,
+  nonEvmScopes: string[] = EXPECTED_NON_EVM_SCOPES,
+) {
+  const expectedNewEvmSessionScopes = evmScopes.map((scope: string) => ({
+    [scope]: getExpectedSessionScope(scope, [
+      multichainAccount.evm.toLowerCase(),
+    ]),
+  }));
+
+  const expectedSolanaSessionScopes = nonEvmScopes.map((scope: string) => {
+    return {
+      [scope]: {
+        accounts: [`${scope}:${multichainAccount.solana}`],
+        methods:
+          scope !== 'solana:4uhcVJyU9pJkvQyS88uRDiswHXSCkY3z' // regnet is missing some methods
+            ? [
+                'signAndSendTransaction',
+                'signTransaction',
+                'signMessage',
+                'signIn',
+                'getGenesisHash',
+                'getLatestBlockhash',
+                'getMinimumBalanceForRentExemption',
+              ]
+            : [
+                'signAndSendTransaction',
+                'signTransaction',
+                'signMessage',
+                'signIn',
+              ],
+        notifications: [],
+      },
+    };
+  });
+
+  for (const expectedSessionScope of [
+    ...expectedNewEvmSessionScopes,
+    ...expectedSolanaSessionScopes,
+  ]) {
+    const [scopeName] = Object.keys(expectedSessionScope);
+    const expectedScopeObject = expectedSessionScope[scopeName];
+    const resultSessionScope = sessionResult.sessionScopes[scopeName];
+
+    assert.deepEqual(
+      expectedScopeObject,
+      resultSessionScope,
+      `${scopeName} does not match expected scope`,
+    );
+  }
+}
+
+export async function connectToMultichainTestDapp(
+  driver: Driver,
+  extensionId: string,
+  multichainAccount: {
+    evm: string;
+    solana: string;
+  } = SECOND_MULTICHAIN_ACCOUNT_IN_SECOND_TEST_E2E,
+  checkPermissions = true,
+) {
+  const testDapp = new TestDappMultichain(driver);
+  await testDapp.openTestDappPage();
+  await testDapp.checkPageIsLoaded();
+  await testDapp.connectExternallyConnectable(extensionId);
+
+  await driver.clickElement('[data-testid="network-checkbox-eip155-1"]');
+  await driver.clickElement(
+    '[data-testid="network-checkbox-solana-5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp"]',
+  );
+
+  await testDapp.clickWalletCreateSessionButton();
+  await driver.switchToWindowWithTitle(WINDOW_TITLES.Dialog);
+  await driver.delay(largeDelayMs);
+
+  const connectAccountConfirmation = new ConnectAccountConfirmation(driver);
+  await connectAccountConfirmation.checkPageIsLoaded();
+  await connectAccountConfirmation.confirmConnect();
+
+  await driver.switchToWindowWithTitle(WINDOW_TITLES.MultichainTestDApp);
+  await testDapp.checkPageIsLoaded();
+
+  if (checkPermissions) {
+    const newgetSessionResult = await testDapp.getSession();
+    validateSessionScopes(newgetSessionResult, multichainAccount);
+  }
+}
+
+export async function checkPermissions(
+  driver: Driver,
+  numberOfNetworks: number,
+) {
+  await driver.switchToWindowWithTitle(WINDOW_TITLES.ExtensionInFullScreenView);
+  const homePage = new HomePage(driver);
+  await homePage.checkPageIsLoaded();
+
+  // change to main wallet view and check permissions
+  // open permissions page and check that the dapp is connected
+  await new HeaderNavbar(driver).openPermissionsPage();
+  const permissionListPage = new PermissionListPage(driver);
+  await permissionListPage.checkPageIsLoaded();
+  await permissionListPage.checkConnectedToSite(DAPP_HOST_ADDRESS);
+  await permissionListPage.checkNumberOfConnectedSites(1);
+  await permissionListPage.openPermissionPageForSite(DAPP_HOST_ADDRESS);
+  const sitePermissionPage = new SitePermissionPage(driver);
+  await sitePermissionPage.checkPageIsLoaded(DAPP_HOST_ADDRESS);
+
+  // All chains should be connected.
+  await sitePermissionPage.checkConnectedNetworksNumber(numberOfNetworks);
+}
+
+export async function removePermissions(
+  driver: Driver,
+  chainsToDisconnect: string[],
+) {
+  await driver.switchToWindowWithTitle(WINDOW_TITLES.ExtensionInFullScreenView);
+  const homepage = new HomePage(driver);
+  await homepage.checkPageIsLoaded();
+  await homepage.headerNavbar.openPermissionsPage();
+  const permissionListPage = new PermissionListPage(driver);
+  await permissionListPage.checkPageIsLoaded();
+  await permissionListPage.checkConnectedToSite(DAPP_HOST_ADDRESS);
+  await permissionListPage.checkNumberOfConnectedSites(1);
+  await permissionListPage.openPermissionPageForSite(DAPP_HOST_ADDRESS);
+  const sitePermissionPage = new SitePermissionPage(driver);
+  await sitePermissionPage.checkPageIsLoaded(DAPP_HOST_ADDRESS);
+  await sitePermissionPage.openNetworkPermissionsModal();
+  for (const chainToDisconnect of chainsToDisconnect) {
+    await sitePermissionPage.uncheckNetwork(chainToDisconnect);
+  }
+  await sitePermissionPage.clickConfirmEditNetworksButton();
+}
